"Finishing" touches
Made it so going to .test page goes straight to log in, and home page now has links to navigate the site (browse movies, browse by category). Also just in case added middleware to any routes that lead to changing database entries. Now technically all functionality is present in project. I'll probably keep messing around to refactor/get further acquainted with the framework, but for the purpose of this challenge, it is complete!